Start reorganizing ITensorNetworks module
