+++ src/ITensorNetworks/ITensorNetworks.jl
@@ -3,7 +3,11 @@
 using ITensors
 
 include("models/models.jl")
-include("tensor_networks.jl")
-include("boundary_mps_projectors.jl")
+include("ITensors.jl")
+include("lattices.jl")
+include("inds_network.jl")
+include("itensor_network.jl")
+include("boundary_mps.jl")
+#include("boundary_mps_projectors.jl")
 
 end
